OptimalSlicing

Documentation for OptimalSlicing.

Optimal Slicing find the optimal contraction path and slicing indices given a tensor network.

OptimalSlicing.check_parallel_edges

OptimalSlicing.compare_cost

OptimalSlicing.compare_cost_int

OptimalSlicing.get_all_index

OptimalSlicing.get_bdim

OptimalSlicing.get_best_results

OptimalSlicing.get_index

OptimalSlicing.get_slicing_cost

OptimalSlicing.initialize_table

OptimalSlicing.nested_to_ssa

OptimalSlicing.path_to_0_indexed

OptimalSlicing.search

OptimalSlicing.search

OptimalSlicing.ssa_to_linear

OptimalSlicing.update_intermediate_tensor

OptimalSlicing.update_table

OptimalSlicing.check_parallel_edges — Method

check_parallel_edges(parallel_edges, inds, sliced_inds) -> Bool

Check if the slicedinds are selected in order of paralleledges.

OptimalSlicing.compare_cost — Method

compare_cost(cost1, cost2) -> Int

Compare the cost1 and cost2, which are polynomials. If cost1 < cost2, then return 1. If cost1 == cost2, then return 0. If cost1 > cost2, then return -1.

OptimalSlicing.compare_cost_int — Method

compare_cost_int(cost1, cost2) -> Int

Compare the cost1 and cost2, which are integers. If cost1 < cost2, then return 1. If cost1 == cost2, then return 0. If cost1 > cost2, then return -1.

OptimalSlicing.get_all_index — Method

get_all_index(inputs, T) -> Set

Get the set of all indices that appear in the contraction of the intermediate tensor T.

OptimalSlicing.get_bdim — Method

get_bdim(idx, size_dict) -> Polynomial

Get the bond dimension of the idx.

OptimalSlicing.get_best_results — Method

get_best_results(tn, config, table) -> Tuple

Get the optimal result of contracting the final tensor from the table.

OptimalSlicing.get_index — Method

get_index(inputs, T) -> Set

Get the set of indices of the intermediate tensor T.

OptimalSlicing.get_slicing_cost — Method

get_slicing_cost(tn, config, table, Ta, Tb, Ia, Ib) -> Tuple

Get the cost of the intermediate tensors Ta and Tb with the sliced indices Ia and Ib.

OptimalSlicing.initialize_table — Method

initialize_table(tn, config, table)

Initialize the table for the bfs.

OptimalSlicing.nested_to_ssa — Method

nested_to_ssa(nested_path::Tuple, N::Int) -> Array{Tuple{Int, Int}, 1}

Convert the nested path to the ssa (single static assignment) path.

Example: nestedtossa (((3,), (2,)), ((5,), ((1,), (4,)))) -> [(2, 1), (0, 3), (4, 6), (5, 7)]

OptimalSlicing.path_to_0_indexed — Method

path_to_0_indexed(path::Array{Tuple{Int, Int}, 1}) -> Array{Tuple{Int, Int}, 1}

Convert the path to the 0-indexed path.

OptimalSlicing.search — Method

search(tn, config)

Search the optimal contraction path and slicing of the tensor network.

OptimalSlicing.search — Method

search(tn_name::String)

Search the optimal contraction path and slicing of the tensor network with the given name.

OptimalSlicing.ssa_to_linear — Method

ssa_to_linear(ssa_path::Array{Tuple{Int, Int}, 1}) -> Array{Tuple{Int, Int}, 1}

Convert the ssa (single static assignment) path to the linear path (conventional contraction path).

OptimalSlicing.update_intermediate_tensor — Method

update_intermediate_tensor(tn, config, table, Ta, Tb)

Update the table of intermediate tensor Tab given Ta and Tb.

OptimalSlicing.update_table — Method

update_table(tn, config, table)

Update the table of intermediate tensors.