OptimalSlicing

Documentation for OptimalSlicing.

OptimalSlicing.check_parallel_edges

OptimalSlicing.compare_cost

OptimalSlicing.compare_cost_int

OptimalSlicing.get_all_index

OptimalSlicing.get_bdim

OptimalSlicing.get_best_results

OptimalSlicing.get_index

OptimalSlicing.get_slicing_cost

OptimalSlicing.initialize_table

OptimalSlicing.search

OptimalSlicing.search

OptimalSlicing.update_intermediate_tensor

OptimalSlicing.update_table

OptimalSlicing.check_parallel_edges — Method

check_parallel_edges(parallel_edges, inds, sliced_inds) -> Bool

Check if the slicedinds are selected in order of paralleledges.

OptimalSlicing.compare_cost — Method

compare_cost(cost1, cost2) -> Int

Compare the cost1 and cost2, which are polynomials. If cost1 < cost2, then return 1. If cost1 == cost2, then return 0. If cost1 > cost2, then return -1.

OptimalSlicing.compare_cost_int — Method

compare_cost_int(cost1, cost2) -> Int

Compare the cost1 and cost2, which are integers. If cost1 < cost2, then return 1. If cost1 == cost2, then return 0. If cost1 > cost2, then return -1.

OptimalSlicing.get_all_index — Method

get_all_index(inputs, T) -> Set

Get the set of all indices that appear in the contraction of the intermediate tensor T.

OptimalSlicing.get_bdim — Method

get_bdim(idx, size_dict) -> Polynomial

Get the bond dimension of the idx.

OptimalSlicing.get_best_results — Method

get_best_results(tn, config, table) -> Tuple

Get the optimal result of contracting the final tensor from the table.

OptimalSlicing.get_index — Method

get_index(inputs, T) -> Set

Get the set of indices of the intermediate tensor T.

OptimalSlicing.get_slicing_cost — Method

get_slicing_cost(tn, config, table, Ta, Tb, Ia, Ib) -> Tuple

Get the cost of the intermediate tensors Ta and Tb with the sliced indices Ia and Ib.

OptimalSlicing.initialize_table — Method

initialize_table(tn, config, table)

Initialize the table for the bfs.

OptimalSlicing.search — Method

search(tn, config)

Search the optimal contraction path and slicing of the tensor network.

OptimalSlicing.search — Method

search(tn_name::String)

Search the optimal contraction path and slicing of the tensor network with the given name.

OptimalSlicing.update_intermediate_tensor — Method

update_intermediate_tensor(tn, config, table, Ta, Tb)

Update the table of intermediate tensor Tab given Ta and Tb.

OptimalSlicing.update_table — Method

update_table(tn, config, table)

Update the table of intermediate tensors.